feat(dashboardnavigation): create dashboardnavigation component
